Moved commented out code over to mockup /js/App.js file, put div className in correct spot for li img's to get styles applied.
