+++ src/components/Photo.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 const Photo = props =>
-	<li className="photo-container">
+	<li>
 		<img src={ `https://farm${props.farm}.staticflickr.com/${props.server}/${props.id}_${props.secret}.jpg` } alt="" />
 	</li>
